code rabbit updated

+++ backend/src/controllers/notification.controller.js
@@ -1,7 +1,7 @@
 import asyncHandler from "express-async-handler";
 import { getAuth } from "@clerk/express";
-import {Notification} from "../models/notification.model.js";
-import {User} from "../models/user.model.js";
+import { Notification } from "../models/notification.model.js";
+import { User } from "../models/user.model.js";
 
 export const getNotifications = asyncHandler(async (req, res) => {
   const { userId } = getAuth(req);
@@ -22,9 +22,9 @@ export const deleteNotification = asyncHandler(async (req, res) => {
   const { userId } = getAuth(req);
   const { notificationId } = req.params;
 
-  const user = await User.findOne({ clerkId: userId });
+  if (!userId) return res.status(401).json({ error: "Unauthorized" });
+  const user = await User.findOne({ clerkId: userId }).select("_id");
   if (!user) return res.status(404).json({ error: "User not found" });
-
   const notification = await Notification.findOneAndDelete({
     _id: notificationId,
     to: user._id,
